chore(services): migrate archived services content into services page
