--- src/app/(frontend)/(inner)/archive-services/ServicesROI.tsx
@@ -1,30 +1,0 @@
-export function ServicesROI() {
-  return (
-    <div>
-      <div className="flex max-w-6xl items-start justify-between bg-white px-8 py-12 text-stone-950">
-        <div className="w-[calc(42%_-_12px)] text-4xl font-medium">
-          <p>Design that makes a good impression and investment.</p>
-        </div>
-        <div className="mt-16 w-[calc(50%_-_12px)] text-xl">
-          At Brewww Studio, ROI isnt optional—its essential. We create brand
-          experiences that not only look good but drive real results. Our
-          superior design nurtures relationships and delivers value, proven
-          across industries like fitness, non-profits, media & entertainment,
-          and e-commerce.
-          <div className="mt-12">
-            <p className="mt-5 font-medium text-neutral-800 opacity-60">
-              5.0/5.0 Avg. Rating on{" "}
-              <a className="underline" href="">
-                Facebook
-              </a>
-              {" and "}
-              <a className="underline" href="">
-                Google
-              </a>
-            </p>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
